CDP-37 chore⚙️(lucide): props제어 & 클린코드 & 점검
